Update colRoz.R
I dont think I changed anything in this part as "oz_palette" name is still the same. Only added example codes to help me remember that continuous is an option. I'm trying to figure out if we can add a function to change the hue of the palettes so that it can fit with light-themed background or dark-themed. Ghibli just made 3 versions of the palettes light, med, dark for those themes. But that means there would be 3 times as much palettes to choose.